Fixes #6 (Allow linking to tabs).
Backport of fix from eddies/home-dashi#6:

Adds a new function to app/scripts/main.js that will read the tab based on the
window.location.hash and show it.

I chose to comment out functionality that would send the hash to the address bar
on click: e.g. selecting the sidebar link for "Domestic workers" would update
the location from http://example.com to http://example.com/#domestic. Instead,
the location bar just remains as-is, regardless of which sidebar tab link is
clicked.